backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
web3==6.11.3
python-dotenv==1.0.0
pyjwt==2.8.0
bcrypt==4.1.1